console: Refactor connect in uplink form
