+++ pkg/webui/locales/en.json
@@ -357,6 +357,9 @@
   "console.components.routing-policy-form.index.saveDefaultPolicy": "Save default policy",
   "console.components.routing-policy-form.index.useSpecificPolicy": "Use network specific routing policy",
   "console.components.routing-policy-form.index.doNotUseAPolicy": "Do not use a routing policy for this network",
+  "console.components.uplink-form.index.simulateUplink": "Simulate uplink",
+  "console.components.uplink-form.index.payloadDescription": "The desired payload bytes of the uplink message",
+  "console.components.uplink-form.index.uplinkSuccess": "Uplink sent",
   "console.components.uplink-form.uplink-form.simulateUplink": "Simulate uplink",
   "console.components.uplink-form.uplink-form.payloadDescription": "The desired payload bytes of the uplink message",
   "console.components.uplink-form.uplink-form.uplinkSuccess": "Uplink sent",
@@ -534,6 +537,17 @@
   "console.containers.gateway-location-form.gateway-location-form.locationFromStatusMessage": "Location set automatically from status messages",
   "console.containers.gateway-location-form.gateway-location-form.setLocationManually": "Set location manually",
   "console.containers.gateway-location-form.gateway-location-form.noLocationSetInfo": "This gateway has no location information set",
+  "console.containers.gateway-location-form.index.updateLocationFromStatus": "Update from status messages",
+  "console.containers.gateway-location-form.index.updateLocationFromStatusDescription": "Update the location of this gateway based on incoming status messages",
+  "console.containers.gateway-location-form.index.setGatewayLocation": "Gateway antenna location settings",
+  "console.containers.gateway-location-form.index.locationSource": "Location source",
+  "console.containers.gateway-location-form.index.locationPrivacy": "Location privacy",
+  "console.containers.gateway-location-form.index.placement": "Placement",
+  "console.containers.gateway-location-form.index.indoor": "Indoor",
+  "console.containers.gateway-location-form.index.outdoor": "Outdoor",
+  "console.containers.gateway-location-form.index.locationFromStatusMessage": "Location set automatically from status messages",
+  "console.containers.gateway-location-form.index.setLocationManually": "Set location manually",
+  "console.containers.gateway-location-form.index.noLocationSetInfo": "This gateway has no location information set",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.index.requireAuthenticatedConnectionDescription": "Select which information can be seen by other network participants, including {packetBrokerURL}",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.index.shareGatewayInfoDescription": "Choose this option eg. if your gateway is powered by {loraBasicStationURL}",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.validation-schema.validateEntry": "There must be at least one selected frequency plan ID.",
